Separate class for reader and analyzer
